[BLD-3686] finished demo load
